# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Todo API",
        "description": "This is a sample API",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/checklist": {
            "get": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Get all checklists",
                "operationId": "ba44fc47c06e7adca69f3309915254e2",
                "responses": {
                    "200": {
                        "description": "List of checklists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Checklist"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Create a new checklist",
                "operationId": "9dd662a7a673f10d4177faec0f7deba8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Checklist name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checklist created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checklist/{id}": {
            "get": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Get a checklist by ID",
                "operationId": "d86676c55fb232785f4f3ccc052ca845",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Update a checklist",
                "operationId": "54075345948d0051908470875de1cdc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Updated checklist name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checklist updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Delete a checklist",
                "operationId": "7d4ad43f6dc452f63f8e61d8d5c1910b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist deleted"
                    },
                    "404": {
                        "description": "Checklist not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Checklist": {
                "title": "Checklist",
                "description": "Schema for Checklist model",
                "required": [
                    "user_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the checklist",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID of the user who owns the checklist",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the checklist",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter 'Bearer <token>' to authenticate.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}